Добавляет код по ДЗ 5.14. Модуляция
