+++ js/data.js
@@ -1,0 +1,47 @@
+export const NUMBER_OBJECT_DESCRIPTIONS = 25;
+export const ID_MIN = 1;
+export const ID_MAX = 25;
+export const URL_MIN = 1;
+export const URL_MAX = 25;
+export const LIKES_MIN = 15;
+export const LIKES_MAX = 200;
+
+export const COMMENTS_MIN = 0;
+export const COMMENTS_MAX = 30;
+export const COMMENTS_ID_MIN = 1;
+export const COMMENTS_ID_MAX = COMMENTS_MAX * ID_MAX;
+export const COMMENTS_AVATAR_MIN = 1;
+export const COMMENTS_AVATAR_MAX = 6;
+
+export const NAMES = [
+  'Иван',
+  'Хуан Себастьян',
+  'Мария',
+  'Кристоф',
+  'Виктор',
+  'Юлия',
+  'Люпита',
+  'Вашингтон',
+];
+
+export const MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export const DESCRIPTION_TEXT = [
+  'Вся красота мира в одной картинке.',
+  'Моменты, которые запечатлены навсегда.',
+  'Счастье в каждом кадре.',
+  'Когда слова не нужны, достаточно фотографии.',
+  'История, рассказанная через объектив.',
+  'Остановить время в одном кадре.',
+  'Фотография — это способ улыбнуться в будущем.',
+  'Сегодня — самый лучший день.',
+  'Я не доверяю словам. Я доверяю фотографиям.',
+  'Фотографии — это свидетельство о том, что мы жили.',
+];
